Refactored parser testing
